# openapi.yml
openapi: 3.1.0
info:
  title: SMSGang API
  version: 1.0.0
  description: |
    The SMSGang REST API enables purchasing temporary virtual phone numbers for
    SMS verification on 300+ global platforms. All monetary values are in **Nigerian Naira (₦)**.

    ## Authentication
    The API uses **Laravel Sanctum** bearer tokens. After registering or logging in,
    include the token in every protected request via the `Authorization` header:
    ```
    Authorization: Bearer {token}
    ```

    ## Rate Limits
    | Scope | Limit |
    |---|---|
    | `auth` | 10 req / min |
    | `api` | 60 req / min |
    | `buy` | 5 purchases / min |
    | `sms-check` | 10 checks / min |
    | `webhook` | 30 req / min |

  contact:
    name: SMSGang Support
  license:
    name: Proprietary
servers:
  - url: https://api.smsgang.ng/api
    description: Production
  - url: http://localhost:8000/api
    description: Local Development
...
paths:
...
  /countries/{country}/services:
    get:
      tags: [Catalog]
      summary: Get available services for a country
      operationId: catalog.country.services
      parameters:
        - $ref: '#/components/parameters/CountryCode'
      responses:
        '200':
          description: Services available in the given country
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ServiceCollection'
        '404':
          $ref: '#/components/responses/NotFound'

  # ── ACTIVATIONS ──────────────────────────────
...
  /wallet/balance:
    get:
      tags: [Wallet]
      summary: Get wallet balance
      operationId: wallet.getBalance
      security:
        - BearerToken: []
      responses:
        '200':
          description: Wallet balance retrieved
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/WalletBalance'
        '401':
          $ref: '#/components/responses/Unauthenticated'
...
  /smm/orders:
    post:
      tags: [SMM]
      summary: Create SMM order
      description: Purchase and create a new SMM service order. Deducts cost from wallet.
      operationId: smm.createOrder
      security:
        - BearerToken: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateSmmOrderRequest'
            example:
              smm_service_id: 1
              link: https://www.instagram.com/username
              quantity: 100
      responses:
        '201':
          description: Order created successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SmmOrderResponse'
        '422':
          $ref: '#/components/responses/ValidationError'
        '401':
          $ref: '#/components/responses/Unauthenticated'

    get:
      tags: [SMM]
      summary: List user's SMM orders
      operationId: smm.listOrders
      security:
        - BearerToken: []
      parameters:
        - name: status
          in: query
          schema:
            type: string
            enum: [Pending, In progress, Partial, Completed, Failed, Cancelled]
        - name: per_page
          in: query
          schema:
            type: integer
            default: 20
      responses:
        '200':
          description: User's SMM orders
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SmmOrderList'
        '401':
          $ref: '#/components/responses/Unauthenticated'
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: Laravel Sanctum personal access token

  # ── PARAMETERS ──────────────────────────────
  parameters:
...
    CountryCode:
      name: country
      in: path
      required: true
      schema:
        type: string
      description: Country ISO code or slug (e.g. `NG`)
...
  schemas:
...
    Service:
      type: object
      properties:
        id:
          type: integer
          example: 42
        name:
          type: string
          example: WhatsApp
        slug:
          type: string
          example: whatsapp
        is_active:
          type: boolean
          example: true
        created_at:
          type: string
          format: date-time
...
    MessageResponse:
      type: object
      properties:
        message:
          type: string
    ValidationErrorResponse:
      type: object
      properties:
        message:
          type: string
          example: The given data was invalid.
        errors:
          type: object
          additionalProperties:
            type: array
            items:
              type: string
          example:
            email:
              - The email has already been taken.

    # ---- Paginated Collections ----------------
...
    ServiceCollection:
      type: object
      properties:
        data:
          type: array
          items:
            $ref: '#/components/schemas/Service'
...
    WalletBalance:
      type: object
      properties:
        balance:
          type: string
          format: decimal
          example: "150000.00"
        currency:
          type: string
          example: NGN
...
    CreateSmmOrderRequest:
      type: object
      required: [smm_service_id, link, quantity]
      properties:
        smm_service_id:
          type: integer
          example: 1
        link:
          type: string
          format: url
          example: https://www.instagram.com/username
        quantity:
          type: integer
          minimum: 1
          example: 100
        runs:
          type: integer
          nullable: true
          example: null
        interval:
          type: integer
          nullable: true
          example: null
        comments:
          type: string
          nullable: true
          example: null
    SmmOrderResponse:
      type: object
      properties:
        message:
          type: string
          example: Order created successfully.
        order:
          type: object
          properties:
            id:
              type: integer
            crestpanel_order_id:
              type: string
            service_name:
              type: string
            link:
              type: string
            quantity:
              type: integer
            total_cost_ngn:
              type: string
              format: decimal
            status:
              type: string
            created_at:
              type: string
              format: date-time
        remaining_balance:
          type: string
          format: decimal
    SmmOrderList:
      type: object
      properties:
        data:
          type: array
          items:
            type: object
            properties:
              id:
                type: integer
              crestpanel_order_id:
                type: string
              service_name:
                type: string
              link:
                type: string
              quantity:
                type: integer
              total_cost_ngn:
                type: string
                format: decimal
              status:
                type: string
              created_at:
                type: string
                format: date-time
        pagination:
          $ref: '#/components/schemas/PaginationData'
...
  responses:
    Unauthenticated:
      description: Unauthenticated — missing or invalid bearer token
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/MessageResponse'
          example:
            message: Unauthenticated.
...
    NotFound:
      description: Resource not found
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/MessageResponse'
          example:
            message: Not found.
    ValidationError:
      description: Validation failed
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ValidationErrorResponse'